chore: update test and fix cyclic deps
